rename page-path-utils to path-utils
