+++ src/client/js/legacy/crowi.js
@@ -7,7 +7,7 @@ import ReactDOM from 'react-dom';
 
 import { debounce } from 'throttle-debounce';
 
-const pagePathUtils = require('@commons/util/page-path-utils');
+const pathUtils = require('@commons/util/path-utils');
 const entities = require('entities');
 const escapeStringRegexp = require('escape-string-regexp');
 require('jquery.cookie');
@@ -343,7 +343,7 @@ $(function() {
     if (name.match(/.+\/$/)) {
       name = name.substr(0, name.length - 1);
     }
-    top.location.href = pagePathUtils.encodePagePath(name) + '#edit';
+    top.location.href = pathUtils.encodePagePath(name) + '#edit';
     return false;
   });
